Use a Svelte attachment for the demo and delete what that orphaned
The dropzone demo is an attachment rather than an action, and the site
now demos the library from this workspace instead of a published 5.9.3.
That also fixes the first-request 500 in `vite dev`: the old action
dynamically imported dropzone during server rendering, which left
Svelte's dev-only element validation reading a null SSR context. It was
never a warm-up quirk, as the README claimed -- that note is gone.

The tracking actions, PricingPlans and the /plus route went with it, and
walking the import graph from the routes turned up four more files that
nothing could reach: Feature, FeatureSet, breakpoints.ts, and the
GoatCounter event helper, whose only callers were those actions.
Pageview tracking is unaffected; it comes from the script in app.html.

StyledLink kept a `use` prop that existed solely to attach those
tracking actions. No caller passed it and its type import no longer
resolved. The layout also preloaded the /plus backdrop, which no page
can show any more.

Two things had to follow from depending on the workspace library. The
website resolves it through dist/, so the site build now builds the
library first -- without that, a clean checkout fails to resolve
"dropzone". And the stylesheet comes from the workspace rather than a
pinned unpkg URL, which would have drifted from the code it styles the
moment the library version moved.

Left alone: 1.8 MB of images under static/ that nothing references,
including a "js copy.jpg". Those are assets rather than components.
