+++ apps/website/package.json
@@ -23,7 +23,7 @@
     "@sveltejs/kit": "3.0.0-next.27",
     "@sveltejs/vite-plugin-svelte": "^7.3.0",
     "@types/dropzone": "^5.7.9",
-    "dropzone": "5.9.3",
+    "dropzone": "workspace:*",
     "postcss": "^8.5.6",
     "postcss-functions": "^4.0.2",
     "svelte": "^5.57.0",
