+++ apps/website/src/app.html
@@ -16,8 +16,6 @@
 
     %sveltekit.head%
 
-    <link href="https://unpkg.com/dropzone@5.9.3/dist/min/dropzone.min.css" rel="stylesheet" />
-
     <script
       data-goatcounter="https://dropzone.goatcounter.com/count"
       async
