--- apps/website/src/lib/FeatureSet.svelte
@@ -1,15 +1,0 @@
-<script lang="ts">
-  import type { Snippet } from "svelte";
-
-  let { children }: { children: Snippet } = $props();
-</script>
-
-<div>{@render children()}</div>
-
-<style>
-  div {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(14rem, 1fr));
-    grid-gap: 1.5rem;
-  }
-</style>
