--- apps/website/src/routes/plus/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-  import Intro from "#lib/Intro.svelte";
-</script>
-
-<Intro title="Dropzone Plus has been discontinued">
-  <p>We're sorry to inform you that we have made the hard decision to discontinue Dropzone Plus.</p>
-  <p>All active subscriptions have been refunded.</p>
-</Intro>
